backend/db/connection.py
import os
import psycopg2
from contextlib import contextmanager
from dotenv import load_dotenv
from pathlib import Path

dotenv_path = Path(__file__).parent.parent.parent / ".env"
load_dotenv(dotenv_path)

@contextmanager
def get_connection():
    conn = psycopg2.connect(
        host=os.getenv("DB_HOST"),
        port=os.getenv("DB_PORT"),
        database=os.getenv("DB_NAME"),
        user=os.getenv("DB_USER"),
        password=os.getenv("DB_PASSWORD"),
    )
    try:
        yield conn
    finally:
        conn.close()
