backend/requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argostranslate==1.10.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.0
cachetools==6.2.4
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
conan==2.20.1
confection==0.1.5
ctranslate2==4.6.0
cymem==2.0.11
emoji==2.15.0
fastapi==0.128.0
fasteners==0.20
filelock==3.20.0
fsspec==2025.9.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
narwhals==2.15.0
networkx==3.5
numpy==2.3.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
patch-ng==1.18.1
pillow==12.1.0
preshed==3.0.10
protobuf==6.33.0
psycopg2-binary==2.9.11
pyarrow==22.0.0
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
sacremoses==0.1.1
selenium==4.39.0
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.4.1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.45
srsly==2.5.1
stanza==1.10.1
starlette==0.50.0
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
toml==0.10.2
torch==2.9.0
tornado==6.5.4
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
uvicorn==0.40.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.9.0
wrapt==2.0.0
wsproto==1.3.2
xx_sent_ud_sm @ https://github.com/explosion/spacy-models/releases/download/xx_sent_ud_sm-3.8.0/xx_sent_ud_sm-3.8.0-py3-none-any.whl#sha256=1b5fb9321723b395f76458d2a13b7a3673c537eb58c0a9df3778f4e0950180b4
apscheduler==3.10.4

